% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compareSeries}
\alias{compareSeries}
\title{Title}
\usage{
compareSeries(expected, actual)
}
\arguments{
\item{expected}{}

\item{actual}{}
}
\value{
a summary of some sort
}
\description{
Description
}
\examples{
\dontrun{
library(verifyImpl)
}
}

