% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extractSbS}
\alias{extractSbS}
\title{Extract data side by side from two time series}
\usage{
extractSbS(expected, actual, varId, tSpan = NA)
}
\arguments{
\item{expected}{a time series (e.g. xts). Suggests to use this for the expected or observed data set}

\item{actual}{a time series (e.g. xts). Suggests to use this for the actual data set that should match the expected one.}

\item{varId}{a character vector of length 1, the name of the series to identify}

\item{tSpan}{a vector of length 2, coercible to a POSIXct, or NA (default value) to not temporally subset the returned time series.}
}
\value{
series with renamed column names
}
\description{
Extract data side by side from two time series
}
\examples{
\dontrun{
library(verifyImpl)
}
}

