% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotCompare}
\alias{plotCompare}
\title{Extract and plot data side by side from two time series}
\usage{
plotCompare(expected, actual, varId, tSpan = NA,
  title = paste("before/after state trace", varId))
}
\arguments{
\item{expected}{a time series (e.g. xts). Suggests to use this for the expected or observed data set}

\item{actual}{a time series (e.g. xts). Suggests to use this for the actual data set that should match the expected one.}

\item{varId}{a character vector of length 1, the name of the series to identify}

\item{tSpan}{a vector of length 2, coercible to a POSIXct, or NA (default value) to not temporally subset the returned time series.}

\item{title}{title to give to the resulting plot}
}
\value{
a plot object
}
\description{
Extract and plot data side by side from two time series
}
\examples{
\dontrun{
library(verifyImpl)
}
}

