% Generated by roxygen2 (4.0.2): do not edit by hand
\name{renameSeries}
\alias{renameSeries}
\title{Rename the column names of a data frame, using a lookup table}
\usage{
renameSeries(series, namesLookup)
}
\arguments{
\item{series}{the data frame with column names to rename.}

\item{namesLookup}{a list, lookup table with names and values as characters}
}
\value{
series with renamed column names
}
\description{
Rename the column names of a data frame, using a lookup table
}
\examples{
\dontrun{
library(verifyImpl)
}
}

