% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{verifyImpl-package}
\alias{verifyImpl-package}
\title{Tools to verify the correct implementation of model simulation}
\description{
Tools to verify the correct implementation of model simulation
}
\details{
\tabular{ll}{
Package: \tab verifyImpl\cr
Type: \tab Package\cr
Version: \tab 0.5-1\cr
Date: \tab 2014-12-23\cr
URL: \tab https://github.com/jmp75/verifyImpl \cr
Release Notes: \tab Refactor functionalities not specific to SWIFT .\cr
License: \tab GPL 3\cr
}
}
\author{
Jean-Michel Perraud \email{jean-michel.perraud_at_csiro.au}
}
\keyword{Side-by-side,}
\keyword{testing}
\keyword{verification,}

